Rename ComponentsProvider to ContextSystemProvider
