+++ packages/context/src/index.js
@@ -1,6 +1,8 @@
-export * from './ComponentsProvider';
 export * from './ContextSystemProvider';
 export * from './connect';
 export * from './useContextSystem';
 export * from './createStore';
 export * from './utils';
+
+// Deprecated. use ContextSystemProvider instead.
+export * from './ComponentsProvider';
